[hotfix][yarn] Move yarn/Utils#getEnvironmentVariables to BootstrapTools
The method BootstrapTools#getEnvironmentVariables could be reused by Yarn and Kubernetes.
