+++ flink-yarn-tests/src/test/java/org/apache/flink/yarn/UtilsTest.java
@@ -81,29 +81,6 @@ public class UtilsTest extends TestLogger {
 	}
 
 	@Test
-	public void testGetEnvironmentVariables() {
-		Configuration testConf = new Configuration();
-		testConf.setString("yarn.application-master.env.LD_LIBRARY_PATH", "/usr/lib/native");
-
-		Map<String, String> res = Utils.getEnvironmentVariables("yarn.application-master.env.", testConf);
-
-		Assert.assertEquals(1, res.size());
-		Map.Entry<String, String> entry = res.entrySet().iterator().next();
-		Assert.assertEquals("LD_LIBRARY_PATH", entry.getKey());
-		Assert.assertEquals("/usr/lib/native", entry.getValue());
-	}
-
-	@Test
-	public void testGetEnvironmentVariablesErroneous() {
-		Configuration testConf = new Configuration();
-		testConf.setString("yarn.application-master.env.", "/usr/lib/native");
-
-		Map<String, String> res = Utils.getEnvironmentVariables("yarn.application-master.env.", testConf);
-
-		Assert.assertEquals(0, res.size());
-	}
-
-	@Test
 	public void testCreateTaskExecutorCredentials() throws Exception {
 		File root = temporaryFolder.getRoot();
 		File home = new File(root, "home");
